Eric Sunshine made the following 64-bit platform compatibility changes:
-*- Converted AWS to deal with intptr_t rather than void* for all of its
    opaque data needs. iAwsSinkManager::CreateSink(),
    iAwsComponent::GetProperty(), iAwsComponent::SetProperty(),
    iAwsSink::RegisterTrigger(), iAwsParmList::AddOpaque(),
    iAwsParmList::GetOpaque(), and related callbacks all now expect
    intptr_t.  Since converting client AWS code to the new API can be
    rather onerous on account of the amount of repetitious manual labor
    involved, the old void* API is still present, though heavily
    deprecated. As an aid to clients in upgrading their code to the new
    API, one of two control macros can be defined.  Defining
    AWS_OBSOLETE_VOIDP removes from the API all methods dealing with
    void*'s, thus invocations of any such methods will generate
    compilation errors. Alternately, defining AWS_DEPRECATE_VOIDP, will
    cause invocations of methods dealing with void*'s to generate
    compilation warnings when using compilers which support deprecation
    warnings.

-*- Converted the event system to deal with intptr_t rather than void* for
    the opaque `info' member of csevCommand-type events.
    iEventOutlet::Broadcast(), iEventOutlet::ImmediateBroadcast(),
    csEvent::csEvent(), and csEventCommandData::Info all now expect
    intptr_t.

-*- Converted CSWS to deal with intptr_t rather than void* for all of its
    opaque data needs. csComponent::SendCommand(),
    csComponent::SendBroadcast(), csComponent::ForEach(),
    csKeyboardAccelerator::Command(), csTimer::csTimer(),
    csKeyboardAccelerator::Broadcast(), csListBoxItem::ForEachItem(),
    csTreeItem::ForEachItem(), csTreeBox::ForEachItem(), and the related
    callbacks all now expect intptr_t.

-*- Fixed a number of additional potential 64-bit problems reported by the
    compiler throughout the project.


git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@20598 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ apps/tests/gfxtst/gfxtest.cpp
@@ -294,8 +294,8 @@ static bool process_file (const char *fname)
 
   if (opt.verbose || opt.info)
   {
-    printf ("Image size: %d x %d pixels, %d bytes\n", ifile->GetWidth (),
-      ifile->GetHeight (), csImageTools::ComputeDataSize (ifile));
+    printf ("Image size: %d x %d pixels, %lu bytes\n", ifile->GetWidth (),
+      ifile->GetHeight (), (unsigned long)csImageTools::ComputeDataSize(ifile));
     int fmt = ifile->GetFormat ();
     printf ("Image format: %s, alpha channel: %s\n",
       (fmt & CS_IMGFMT_MASK) == CS_IMGFMT_NONE ? "none" :
